Move Castle-only tests to the right place
